moved 'image_to_tensor' to loggers.functional
